Week 3 ratelimitter done
The most intersting thing was the work with objects in JS, like how you can access them and the work with headers was
interesting aswell, at first I wasn't getting it. The most confusing bit is the || operator with the value assignment for the
object since that is what gives the value to the object.